Split runner out into module
